YOLO/requirements.txt
streamlit
torch
torchvision
ultralytics
numpy
Pillow
opencv-python-headless